Refactor Utils making a BlefActivity class
